+++ src/app/manage/media/media-overview/media-overview.component.ts
@@ -100,7 +100,7 @@ export class MediaOverviewComponent implements OnInit, OnDestroy{
     });
 
     if (result) {
-      await this.service.deleteClip(clipId);
+      await this.service.deleteAction(clipId);
     }
   }