+++ src/app/manage/triggers/events-overview/twitch-event-info/timed-event-info.component.ts
@@ -1,9 +1,8 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {TimedClip, TwitchTypesArray} from '@memebox/contracts';
-import {AppQueries} from '../../../../../../projects/app-state/src/lib/state/app.queries';
+import {TimedAction, TwitchTypesArray} from '@memebox/contracts';
+import {AppQueries, AppService} from '@memebox/app-state';
 import {map} from 'rxjs/operators';
 import {MatCheckboxChange} from '@angular/material/checkbox';
-import {AppService} from '../../../../../../projects/app-state/src/lib/state/app.service';
 
 @Component({
   selector: 'app-timed-info',
@@ -13,7 +12,7 @@ import {AppService} from '../../../../../../projects/app-state/src/lib/state/app
 export class TimedEventInfoComponent implements OnInit {
   twitchEvents = TwitchTypesArray;
   @Input()
-  item: TimedClip;
+  item: TimedAction;
 
 
   allInformations$ = this.appQueries.state$.pipe(